Implementei a biblioteca de comunicacao padrao da Microchip para o protocolo SPI, gerada pelo MCC
